scripts/embeddings/requirements-gcp.txt
# Google Cloud Storage
google-cloud-storage>=2.10.0

# Machine Learning & Embeddings
sentence-transformers>=2.2.0
hnswlib>=0.7.0
numpy>=1.24.0

# Nostr Protocol
websockets>=12.0
secp256k1>=0.14.0

# Additional dependencies
torch>=2.0.0
transformers>=4.30.0
